Use separate jobs for each task (#180)
* Use separate jobs for each task

Now the CI workflow runs on every pull request and every push to the
main branch. This runs ESLint, Prettier, tsd, and Vitest, all in their
own job.

The Integration tests workflow runs on the main branch. It runs all
tests, publishes a coverage badge, and sends a Slack notification on
failure.

* Restore meaningful GitHub action step names

* Revert unwanted change