Sync alphalib 2025 05 06 (#226)
* w

* w

* Update template.ts

* Fix deep types error (#227)

* Fix deep types error

One of the big features of TypeScript, is that it can infer types based
on context. This is powerful, but it comes at a cost. Determining types
based on inference consumes much more resources than using a type
annotation. This is an important reason to add explicit type
annotations.

We heavily rely on Zod though. Zod is powerful, but it heavily relies on
inference. This means that by using Zod, we sacrifice type checking
performance. Zod 4 will supposedly be more performant. This is great, but
by definition it can’t be as performant as regular types.

When the types become really complex, TypeScript may error on this. An
explicit type annotation in the right spot fixes this. This PR uses this
to resolve the type error. I can’t explain why this surfaced here, but
not in our internal repos these schemas originate from.

* Update src/alphalib/types/template.ts

---------

Co-authored-by: Kevin van Zonneveld <vanzonneveld@gmail.com>

---------

Co-authored-by: Remco Haszing <remcohaszing@gmail.com>